* p4a-packages:
  Validate new PyPS unfold method
  Add unfolding, pyps made ! Just great :)
  use new function instead of macro definition.
  improve assert string.
  Don't disable warning on hash_put globally, instead use overwrite_ordering_of_the_statement_to_current_mapping() from time to time
  Add overwrite_ordering_of_the_statement_to_current_mapping()
  add hash_overwrite for MA
  Don't warn on hash_put redefinition in loop fusion
  Add results to new tests
  Copy tests from Pointers and add new tests to illustrate the need of interprocedural analysis.
  Add Makefile.in to svn:ignore property
  Casting from pointer to integer (bool) is forbidden, use intptr_t instead
  Cast to intptr_t to avoid warning when using booleans
  Add new examples to illustrate the use of the property ALIASING_ACROSS_TYPES and it's effect on gloabl variables. Sinks for global variables are now in demand.
  Revert 20186: casting bool as integer
  Revert previous commit for bool size.
  Revert previous commit to int for default bool size.

+++ packages/PIPS/pips/src/Libs/c_syntax/util.c
@@ -961,10 +961,10 @@ entity FindOrCreateCurrentEntity(string name,
       pips_assert("block_scope is a block scope", string_block_scope_p(block_scope));
     }
     pips_debug(5,"type %p: %s\n", ct, list_to_string(safe_c_words_entity(ct, NIL)));
-    pips_debug(5,"is_typedef: %d\n",(int)is_typedef);
-    pips_debug(5,"is_static: %d\n",(int)is_static);
-    pips_debug(5,"is_external: %d\n",(int)is_external);
-    pips_debug(5,"is_formal: %d\n",(int)is_formal);
+    pips_debug(5,"is_typedef: %d\n",is_typedef);
+    pips_debug(5,"is_static: %d\n",is_static);
+    pips_debug(5,"is_external: %d\n",is_external);
+    pips_debug(5,"is_formal: %d\n",is_formal);
     if (is_formal)
       pips_debug(5,"of current function %s\n",entity_user_name(function));
     /* function is only used for formal variables*/
