+++ packages/PIPS/pips/src/Libs/step/directive_do.c
@@ -187,7 +187,7 @@ bool is_begin_directive_omp_parallel_do(directive __attribute__ ((unused)) curre
 
   b = type_directive_omp_parallel_do_p(directive_type(next));
 
-  pips_debug(1,"b = %d\n", (int)b);
+  pips_debug(1,"b = %d\n", b);
   return b;
 }
 
@@ -216,7 +216,7 @@ bool is_end_directive_omp_end_parallel_do(directive current, directive next)
   b1 = type_directive_omp_parallel_do_p(directive_type(current));
   b2 = type_directive_omp_end_parallel_do_p(directive_type(next));
 
-  pips_debug(1,"b1 = %d, b2 = %d\n", (int)b1, (int)b2);
+  pips_debug(1,"b1 = %d, b2 = %d\n", b1, b2);
   return b1 && b2;
 }
 
@@ -277,7 +277,7 @@ bool is_begin_directive_omp_do(directive __attribute__ ((unused)) current, direc
 
   b = type_directive_omp_do_p(directive_type(next));
 
-  pips_debug(1,"b = %d\n", (int)b);
+  pips_debug(1,"b = %d\n", b);
   return b;
 }
 
@@ -304,7 +304,7 @@ bool is_end_directive_omp_end_do(directive current, directive next)
   b1 = type_directive_omp_end_do_p(directive_type(next));
   b2 = type_directive_omp_do_p(directive_type(current));
 
-  pips_debug(1,"b1 = %d, b2 = %d\n", (int)b1, (int)b2);
+  pips_debug(1,"b1 = %d, b2 = %d\n", b1, b2);
   return b1 && b2;
 }
 
@@ -402,7 +402,7 @@ bool update_label_do_directive_module_name(directive begin, directive end)
 
 string directive_omp_do_to_string(directive d,bool close)
 {
-  pips_debug(1, "d=%p, close=%u\n",d,(int)close);
+  pips_debug(1, "d=%p, close=%u\n",d,close);
   if (close)
     return strdup(END_DO_TXT);
   else
@@ -411,7 +411,7 @@ string directive_omp_do_to_string(directive d,bool close)
 
 string directive_omp_parallel_do_to_string(directive d,bool close)
 {
-  pips_debug(1, "d=%p, close=%u\n",d,(int)close);
+  pips_debug(1, "d=%p, close=%u\n",d,close);
   if (close)
     return strdup(END_PARALLEL_DO_TXT);
   else
